+++ document_text_extraction/text_extraction/image_text_extraction.py
@@ -1,0 +1,1 @@
+"""Image(PNG, JPG, Tiff) Text Extraction class."""