+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased
 
 
+- feat(text_extraction/): Se agregó módulo text_extraction
 - Initial commit